+++ frontend/src/app/hair-dresser/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { HairDresserDetail } from "@/features/hair-dresser-detail";
+import { HairDresserDetail } from "@/features/HairDresserDetail";
 
 export default async function Home() {
 	return (
