+++ frontend/src/features/HairDresserDetail/components/ContentTabs.tsx
