+++ frontend/src/features/HairDresserDetail/components/PhotoCard.tsx
