+++ frontend/src/features/HairDresserDetail/components/index.ts
@@ -2,3 +2,4 @@ export { ContentTabs } from "./ContentTabs";
 export { ProfileCard } from "./ProfileCard";
 export { PhotoCard } from "./PhotoCard";
 export { ReviewsCard } from "./ReviewsCard";
+export { HairDresserUserInfo } from "./HairDresserUserInfo";
